+++ nextgen/codegen/src/codegen_error.rs
@@ -5,6 +5,13 @@ use thiserror::Error;
 
 #[derive(Error, Debug, Diagnostic)]
 pub enum CodegenError {
+    #[diagnostic(code(codegen::template::fs_only))]
+    #[error(
+        "Unable to create a new template, as the destination must be a local file system path.\nPlease add a file path to the {} setting.",
+        "generator.templates".style(Style::Property),
+    )]
+    CreateFileSystemOnly,
+
     #[diagnostic(code(codegen::template::exists))]
     #[error(
         "A template with the name {} already exists at {}.",